requirements.txt
### requirements.txt

python-telegram-bot==21.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0